+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 
 ## Unreleased
 
++ Add longer aliases for column types in DDL, e.g. `INT64`, `TEXT`.
+
 * Better schema change detection in `ENUM` columns - independent of whitespaces used between values.